migrate SSH layer to frittura-ssh-core
