--- src/ssh/mod.rs
@@ -1,7 +1,0 @@
-mod channel;
-mod client;
-mod server;
-mod utils;
-
-pub use crate::ssh::channel::SSHWriterProxy;
-pub use crate::ssh::server::AppServer;
